changed route names, added eventId & params to meet and tournament
