+++ app/meet/eventId/page.tsx
@@ -1,25 +1,22 @@
 import { EditButton } from "@/components/EditButton";
 
-const placeHolder = [
-  "Public",
-  "Pros only",
-  "One on One",
-  "4h",
-  "16 players",
-  "knockout",
-];
+const placeHolder = ["Public", "Beginner friendly", "One on One", "1.5 h"];
 
-export default function TournamentDetail() {
+export default function MeetDetail({
+  params,
+}: {
+  params: { eventId: string };
+}) {
+  const { eventId } = params;
   return (
     <main className="h-screen w-screen p-4">
-      <EditButton />
       {/* Header */}
-      <div className="flex flex-row justify-end pb-8 w-full h-6"></div>
+      <EditButton />
       {/* main */}
       <section>
         {/* Hero */}
         <div className="flex flex-col justify-center items-center text-center gap-4">
-          <h1 className="text-4xl font-bold">Tournament</h1>
+          <h1 className="text-4xl font-bold">Meet</h1>
           <h2 className="text-sm border-solid border-zinc-200 border-2 rounded-xl px-16 py-2">
             Erich Zeigner Allee
           </h2>
@@ -35,10 +32,10 @@ export default function TournamentDetail() {
         {/* Tags and Form */}
         <div className="flex flex-col items-stretch justify-center text-center gap-2 py-2 min-h-1/3">
           <h3 className="text-sm border-solid border-zinc-200 border-2 rounded-lg py-2">
-            29.08.2024
+            28.08.2024
           </h3>
           <h3 className="text-sm border-solid border-zinc-200 border-2 rounded-lg py-2">
-            16:00
+            3:00 PM
           </h3>
           {/* Tags */}
           <ul className="flex flex-row flex-wrap gap-2">
@@ -56,7 +53,8 @@ export default function TournamentDetail() {
               Description
             </label>
             <p className="border-zinc-200 border-2 rounded-lg p-2 min-h-20 text-sm">
-              You wanna be the very best? Then join our ultimate tournament!
+              You wann have a fun round of table tennis with us? Then join us!
+              We play friendly&fair.
             </p>
           </div>
         </div>
@@ -64,7 +62,7 @@ export default function TournamentDetail() {
       {/* Button */}
       <div className="flex flex-col items-stretch flex-grow justify-end mb-14 mt-6">
         <button className="bg-black text-white rounded-l py-4">
-          Join Tournament
+          Join Meet
         </button>
       </div>
     </main>
