+++ app/response/[responseId]/page.tsx
@@ -1,23 +1,46 @@
-import { EditButton } from "@/components/EditButton";
+import { Name } from "@/components/name";
+import Link from "next/link";
 
-const placeHolder = ["Public", "Beginner friendly", "One on One", "1.5 h"];
+export default function responsePage({
+  params,
+}: {
+  params: { responseId: string };
+}) {
+  const { responseId } = params;
+  const placeHolder = [
+    "Public",
+    "Pros only",
+    "One on One",
+    "4h",
+    "16 players",
+    "knockout",
+  ];
 
-export default function MeetDetail() {
+  const eventName = "Tournament";
+  const adress = " Erich Zeigner Allee";
+  const date = "29.08.2024";
+  const time = "16:00";
+  const sessionId = "345";
+
   return (
     <main className="h-screen w-screen p-4">
       {/* Header */}
-      <EditButton />
+      <h1 className="text-center font-bold">
+        JeanLuc1312 has invited you to a:
+      </h1>
+
+      <div className="flex flex-row justify-end pb-8 w-full h-6"></div>
       {/* main */}
       <section>
         {/* Hero */}
         <div className="flex flex-col justify-center items-center text-center gap-4">
-          <h1 className="text-4xl font-bold">Meet</h1>
-          <h2 className="text-sm border-solid border-zinc-200 border-2 rounded-xl px-16 py-2">
-            Erich Zeigner Allee
-          </h2>
+          <h2 className="text-4xl font-bold">{eventName}</h2>
+          <h3 className="text-sm border-solid border-zinc-200 border-2 rounded-xl px-16 py-2">
+            {adress}
+          </h3>
           <img
             className="rounded-xl h-2/5 w-full object-fill bg-red-300"
-            src="example.png"
+            src="/example.png"
             alt="pic"
           />
           <button className="bg-zinc-300 text-purple-700 text-xs rounded-full w-1/3 py-1">
@@ -27,10 +50,7 @@ export default function MeetDetail() {
         {/* Tags and Form */}
         <div className="flex flex-col items-stretch justify-center text-center gap-2 py-2 min-h-1/3">
           <h3 className="text-sm border-solid border-zinc-200 border-2 rounded-lg py-2">
-            28.08.2024
-          </h3>
-          <h3 className="text-sm border-solid border-zinc-200 border-2 rounded-lg py-2">
-            3:00 PM
+            {date} - {time}
           </h3>
           {/* Tags */}
           <ul className="flex flex-row flex-wrap gap-2">
@@ -43,21 +63,14 @@ export default function MeetDetail() {
               </li>
             ))}
           </ul>
-          <div className="flex flex-col justify-start text-start gap-2 mt-2">
-            <label htmlFor="description" className="text-sm">
-              Description
-            </label>
-            <p className="border-zinc-200 border-2 rounded-lg p-2 min-h-20 text-sm">
-              You wann have a fun round of table tennis with us? Then join us!
-              We play friendly&fair.
-            </p>
-          </div>
+          <Name />
         </div>
+        <Link href={`/${eventName}-detail/${sessionId}`}></Link>
       </section>
       {/* Button */}
       <div className="flex flex-col items-stretch flex-grow justify-end mb-14 mt-6">
         <button className="bg-black text-white rounded-l py-4">
-          Join Meet
+          Join {eventName}
         </button>
       </div>
     </main>
